+++ ssb_parser/tests/parse_tests.rs
@@ -339,7 +339,7 @@ Texture: Fancy,data,RmFuY3k=
                                 0.0, 0.0, 1.0, 0.0,
                                 0.0, 0.0, 0.0, 1.0
                             ])),
-                            EventObject::TagIdentity,
+                            EventObject::TagReset,
                             EventObject::TagBorder(Border::All(
                                 42.0, 42.0
                             )),